Documentation, Logging, Guilds - finished up the rest of the guild related sample scripts - removed .WithProperties due to issues with sometimes having a string as the 3rd or 4th property causing the wrong method to be called - all .WithProperties converted to .WithProperty chained calls - added documentation for Skills.md and Spells.md - added a couple mapper profiles for some more small things - added configurable world options for LootDropsLockToRewardTargetSecs and GroundItemDespawnTimeMins

+++ Chaos.Schemas/Data/LootDropSchema.cs
@@ -8,12 +8,13 @@ namespace Chaos.Schemas.Data;
 public sealed record LootDropSchema
 {
     /// <summary>
-    ///     A unique id specific to the template of the item that should drop
+    ///     The chance of the item to drop (MUST be between 0 and 1)
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.Never)]
-    public int DropChance { get; set; }
+    public decimal DropChance { get; set; }
+
     /// <summary>
-    ///     The chance of the item to drop
+    ///     A unique id specific to the template of the item that should drop
     /// </summary>
     [JsonRequired]
     public string ItemTemplateKey { get; set; } = null!;
