+++ Chaos/Models/Data/LootDrop.cs
@@ -2,6 +2,6 @@ namespace Chaos.Models.Data;
 
 public sealed record LootDrop
 {
-    public required int DropChance { get; init; }
+    public required decimal DropChance { get; init; }
     public required string ItemTemplateKey { get; init; }
 }
